third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/getComputedStyle-length-unit-expected.txt

PASS testPropertyUnit('visible', 'background-position') is 'px'
PASS testPropertyUnit('visible', 'background-size') is 'px'
PASS testPropertyUnit('visible', '-webkit-border-horizontal-spacing') is 'px'
PASS testPropertyUnit('visible', '-webkit-border-vertical-spacing') is 'px'
PASS testPropertyUnit('visible', 'border-top-width') is 'px'
PASS testPropertyUnit('visible', 'border-right-width') is 'px'
PASS testPropertyUnit('visible', 'border-bottom-width') is 'px'
PASS testPropertyUnit('visible', 'border-left-width') is 'px'
PASS testPropertyUnit('visible', 'border-top-left-radius') is 'px'
PASS testPropertyUnit('visible', 'border-top-right-radius') is 'px'
PASS testPropertyUnit('visible', 'border-bottom-left-radius') is 'px'
PASS testPropertyUnit('visible', 'border-bottom-right-radius') is 'px'
PASS testPropertyUnit('visible', 'outline-width') is 'px'
PASS testPropertyUnit('visible', '-webkit-column-rule-width') is 'px'
PASS testPropertyUnit('visible', '-webkit-column-width') is 'px'
PASS testPropertyUnit('visible', '-webkit-column-gap') is 'px'
PASS testPropertyUnit('visible', '-webkit-mask-position') is 'px'
PASS testPropertyUnit('visible', '-webkit-mask-size') is 'px'
PASS testPropertyUnit('visible', '-webkit-perspective') is 'px'
PASS testPropertyUnit('visible', '-webkit-perspective-origin') is 'px'
PASS testPropertyUnit('visible', '-webkit-text-stroke-width') is 'px'
PASS testPropertyUnit('visible', '-webkit-transform-origin') is 'px'
PASS testPropertyUnit('visible', 'left') is 'px'
PASS testPropertyUnit('visible', 'top') is 'px'
PASS testPropertyUnit('visible', 'right') is 'px'
PASS testPropertyUnit('visible', 'bottom') is 'px'
PASS testPropertyUnit('visible', 'font-size') is 'px'
PASS testPropertyUnit('visible', 'width') is 'px'
PASS testPropertyUnit('visible', 'max-width') is 'px'
PASS testPropertyUnit('visible', 'min-width') is 'px'
PASS testPropertyUnit('visible', 'height') is 'px'
PASS testPropertyUnit('visible', 'max-height') is 'px'
PASS testPropertyUnit('visible', 'min-height') is 'px'
PASS testPropertyUnit('visible', 'letter-spacing') is 'px'
PASS testPropertyUnit('visible', 'word-spacing') is 'px'
PASS testPropertyUnit('visible', 'margin-top') is 'px'
PASS testPropertyUnit('visible', 'margin-right') is 'px'
PASS testPropertyUnit('visible', 'margin-bottom') is 'px'
PASS testPropertyUnit('visible', 'margin-left') is 'px'
PASS testPropertyUnit('visible', 'padding-top') is 'px'
PASS testPropertyUnit('visible', 'padding-right') is 'px'
PASS testPropertyUnit('visible', 'padding-bottom') is 'px'
PASS testPropertyUnit('visible', 'padding-left') is 'px'
PASS testPropertyUnit('visible', 'text-indent') is 'px'
PASS testPropertyUnit('hidden', 'background-position') is 'px'
PASS testPropertyUnit('hidden', 'background-size') is 'px'
PASS testPropertyUnit('hidden', '-webkit-border-horizontal-spacing') is 'px'
PASS testPropertyUnit('hidden', '-webkit-border-vertical-spacing') is 'px'
PASS testPropertyUnit('hidden', 'border-top-width') is 'px'
PASS testPropertyUnit('hidden', 'border-right-width') is 'px'
PASS testPropertyUnit('hidden', 'border-bottom-width') is 'px'
PASS testPropertyUnit('hidden', 'border-left-width') is 'px'
PASS testPropertyUnit('hidden', 'border-top-left-radius') is 'px'
PASS testPropertyUnit('hidden', 'border-top-right-radius') is 'px'
PASS testPropertyUnit('hidden', 'border-bottom-left-radius') is 'px'
PASS testPropertyUnit('hidden', 'border-bottom-right-radius') is 'px'
PASS testPropertyUnit('hidden', 'outline-width') is 'px'
PASS testPropertyUnit('hidden', '-webkit-column-rule-width') is 'px'
PASS testPropertyUnit('hidden', '-webkit-column-width') is 'px'
PASS testPropertyUnit('hidden', '-webkit-column-gap') is 'px'
PASS testPropertyUnit('hidden', '-webkit-mask-position') is 'px'
PASS testPropertyUnit('hidden', '-webkit-mask-size') is 'px'
PASS testPropertyUnit('hidden', '-webkit-perspective') is 'px'
PASS testPropertyUnit('hidden', '-webkit-perspective-origin') is 'px'
PASS testPropertyUnit('hidden', '-webkit-text-stroke-width') is 'px'
PASS testPropertyUnit('hidden', '-webkit-transform-origin') is 'px'
PASS testPropertyUnit('hidden', 'left') is 'px'
PASS testPropertyUnit('hidden', 'top') is 'px'
PASS testPropertyUnit('hidden', 'right') is 'px'
PASS testPropertyUnit('hidden', 'bottom') is 'px'
PASS testPropertyUnit('hidden', 'font-size') is 'px'
PASS testPropertyUnit('hidden', 'width') is 'px'
PASS testPropertyUnit('hidden', 'max-width') is 'px'
PASS testPropertyUnit('hidden', 'min-width') is 'px'
PASS testPropertyUnit('hidden', 'height') is 'px'
PASS testPropertyUnit('hidden', 'max-height') is 'px'
PASS testPropertyUnit('hidden', 'min-height') is 'px'
PASS testPropertyUnit('hidden', 'letter-spacing') is 'px'
PASS testPropertyUnit('hidden', 'word-spacing') is 'px'
PASS testPropertyUnit('hidden', 'margin-top') is 'px'
PASS testPropertyUnit('hidden', 'margin-right') is 'px'
PASS testPropertyUnit('hidden', 'margin-bottom') is 'px'
PASS testPropertyUnit('hidden', 'margin-left') is 'px'
PASS testPropertyUnit('hidden', 'padding-top') is 'px'
PASS testPropertyUnit('hidden', 'padding-right') is 'px'
PASS testPropertyUnit('hidden', 'padding-bottom') is 'px'
PASS testPropertyUnit('hidden', 'padding-left') is 'px'
PASS testPropertyUnit('hidden', 'text-indent') is 'px'
PASS successfullyParsed is true

TEST COMPLETE
